refactor(vtkmath): remove interdependencies between vtkMath and vtkMatrixBuilder
