+++ Sources/tests.js
@@ -1,5 +1,4 @@
 import './Common/Core/LookupTable/test/testCategoricalColors';
-import './Common/Core/Math/test/testMath';
 import './Common/Core/MatrixBuilder/test/testMatrixBuilder';
 import './Common/Core/LookupTable/test/testSetTable';
 import './Common/DataModel/Box/test/testBox';
@@ -76,3 +75,4 @@ import './Testing/testMacro';
 import './Testing/testSerialization';
 import './Testing/testProxy';
 import './Widgets/Widgets3D/ResliceCursorWidget/test/testBoundPlane';
+import './Widgets/Widgets3D/ResliceCursorWidget/test/testRotateVector';
